Add license.txt file for the DMD snap
The license texts in this file are taken from the corresponding license
texts provided in the DMD, druntime and phobos source repositories.  An
overview has been added to summarize the different components, with the
zlib license also included since libphobos statically links to zlib when
it is built.

Note that the DMD compiler license takes into account the announcement
of 7 April 2017 that the backend has been relicensed under the terms of
the Boost Licence (Version 1.0):
https://forum.dlang.org/thread/oc8acc$1ei9$1@digitalmars.com
https://github.com/dlang/dmd/pull/6680

... and in particular, see:
https://forum.dlang.org/post/oc8bie$1gql$1@digitalmars.com