Add snapcraft.yaml defining the snap package and its components
The dmd package uses `classic` confinement to ensure that it can access
the resources of the host system it is installed on, such as gcc, libc
and development libraries against which compiled code needs to link.

The `after:` option has been used to ensure the expected build order of
the different component parts, dmd => druntime => phobos.  When building
druntime and phobos it was necessary to use `make-parameters:` to ensure
that the correct paths were specified in the `DMD`, `DRUNTIME_PATH` and
`VERSION` makefile variables.

Some `organize:` and `stage:` options have been used in order to ensure
a simpler directory layout than is provided by the default install paths
chosen by the various component makefiles.

`AUTO_BOOTSTRAP=1` is used in order to build the D frontend, which needs
`curl` in the build packages in order to download the bootstrap DMD.

Finally, `rdmd`, `dustmite` and `ddemangle` are included from among the
core developer tools.