Add -O3 to the CFLAGS used to build druntime and phobos
This patch manually restores the -O3 flag lost due to the changes made
in commit966f523452610181b8fcc1ab5ff18fb657b749c9, which adds explicit
`CFLAGS` for the druntime and phobos `make-parameters`.  For the reasons
behind this change, see https://github.com/dlang-snaps/dmd.snap/pull/3.

Normally, `posix.mak` starts with a default set of `CFLAGS` which are
then extended with `-g` or `-O3` depending on whether we are making a
debug or release build.  However, if we explicitly specify `CFLAGS` when
invoking the makefile, `posix.mak` takes this as a hard override and no
longer adds in the `-O3` option.  We therefore need to manually include
it in our custom `CFLAGS`.