Update to 2.076.0 stable release
This patch brings the packaged dmd, druntime, phobos and tools up to
date with the latest stable release.

The v2.076.0 release includes some changes to how the library and tools
makefiles select the DMD with which to build: instead of the old `DMD`
variable pointing to the compiler executable, a new `DMD_DIR` variable
pointing to the compiler location is expected/preferred.  The affected
parts' `make-parameters` have been updated accordingly, including the
addition of correct `DRUNTIME_PATH` and `PHOBOS_PATH` values for the
`tools` part.

The `dmd` part's `prepare:` script has also been removed: `git describe`
output is now preferred over the contents of the `VERSION` file, so the
latter no longer needs to be overwritten.  (The latter is now checked to
be correct by the D release process, so should itself also be reliable.)